Fix some errors not being rendered to GitHub annotations
